Fix type

+++ src/config.rs
@@ -11,5 +11,5 @@ pub const CAMERA_DELTA: f32 = 10.0;
 pub const GRID_COLOR: Color = Color::RGB(64, 64, 64);
 pub const CELL_COLOR: Color = Color::RGB(0, 255, 0);
 pub const FEEDBACK_COLOR: Color = Color::RGB(255, 255, 255);
-pub const QT_DIM: u32 = 15;
+pub const QT_DIM: usize = 15;
 pub const STEP: usize = 1;
